Tests/LibWeb/Text/expected/UIEvents/pointermove-platform-data.txt

pointermove trusted=true pointerType=mouse isPrimary=true client=(25, 35) button=0 buttons=0 view=true
